feat(lib): initial test for Requests
This is monumental, with this boour.hs has proven to be more than a concept.
More tests are due but effectively this marks a huuge success for the project.
I am quite pleased with my own work.
Now we keep at at it and write more tests and lock everything donw :D

Milestone: OUT OF POC PHASE